fix: add useStatusStore, split queries into submodules, fix listeners to use separate status store
