+++ src/core-store/listeners/index.ts
@@ -1,13 +1,16 @@
 import { listen } from '@tauri-apps/api/event';
 
 import type { ArrowListItem } from '@/domain/arrow';
+import type { ConnectionStatus } from '@/domain/connection';
 
 import { useArrowStore } from '../store/arrows';
-import type { CoreStatus } from '../store/arrows';
+import { useStatusStore } from '../store/status';
+import type { RuntimeUpdateDTO } from '../dtos/v0/runtime';
+import { toRuntimeUpdate } from '../dtos/v0/runtime';
 
 export async function setupListeners(): Promise<void> {
-	await listen<{ status: CoreStatus }>('core://status', (e) => {
-		useArrowStore.getState().setStatus(e.payload.status);
+	await listen<{ status: ConnectionStatus }>('core://status', (e) => {
+		useStatusStore.getState().setStatus(e.payload.status);
 		if (e.payload.status === 'starting') {
 			useArrowStore.getState().resetArrows();
 		}
@@ -21,12 +24,8 @@ export async function setupListeners(): Promise<void> {
 		useArrowStore.getState().removeArrow(e.payload.namespace);
 	});
 
-	await listen<{
-		namespace: string;
-		state: ArrowListItem['state'];
-		active_run: ArrowListItem['active_run'];
-		last_outcome: ArrowListItem['last_outcome'];
-	}>('runtime://update', (e) => {
-		useArrowStore.getState().applyRuntimeUpdate(e.payload);
+	await listen<RuntimeUpdateDTO>('runtime://update', (e) => {
+		const update = toRuntimeUpdate(e.payload);
+		useArrowStore.getState().applyRuntimeUpdate(update);
 	});
 }
