+++ src/core-store/store/index.ts
@@ -1,2 +1,2 @@
 export { useArrowStore } from './arrows';
-export type { CoreStatus } from './arrows';
+export { useStatusStore } from './status';
